Bug Fixes
Fixed issue with not being in a group not wiping the list. Had to move to a simple timer to move time along a bit since there is no good event to know when it is safe to test for in group (and get the right anwser)

Revamp initialization of network when in a contested race condition from back off scheme to deterministic netlist ordering.

Introduced network protocol versioning, which will help in the future to avoid issues with differing versions.

Made the (inevitable) globals less likely to clash by ensuring all are prefixed by SteaSummon.