Bug fix, a few net changes
Made network code a little more idempotent.

Added replay of events that occur during init but the state is not there to act on. Removed timer to wait for a leader to emerge, since replay makes that bodge unecessary.

Added more protective code around frames that were assumed to always exist, but it turns out asynchronicity has a way of changing that.

Pruned the raid code of all the info collection it turned out was not needed or there was a better way to get at it.

Fixed protocol around destination and at destination to maintain the at dest list properly, there was also a bug around

Caught some places that operated on the waiting table directly and funnelled them through existing methods, or created them.

Made the disconnect message clear that its disconnecting the client from the addon network.

Simplified protocol around destination now that more rigorous checks are made before reporting.